requirements/base.txt
django-jazzmin==3.0.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.10
dj-database-url==2.3.0
redis==5.1.1
django-redis==5.4.0
celery[redis]==5.4.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
Pillow==11.0.0
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.7.0
charset-normalizer==3.3.2